+++ reactapp/src/components/Dropdown/DropDown.js
@@ -1,8 +1,9 @@
 // Dependencies
-import React from "react";
+import React, { Fragment } from "react";
 // import { Route } from "react-router-dom";
 // import { connect } from "react-redux";
 import styled from 'styled-components';
+import axios from 'axios';
 
 // styled-components
 const DropDownWrapper = styled.div`
@@ -24,14 +25,63 @@ const LogInLink = styled.a`
 const SignUpLink = styled.a`
   font-size: 14px;
 `;
+
+const WelcomeMessage = styled.p`
+  font-size: 14px;
+  margin-right: 18px;
+`;
+
+const SignOutLink = styled.a`
+  font-size: 14px;
+`;
 const DropDown = props => {
+
+  function getUserInfo() {
+    axios.get('http://localhost:5000/api/users/user', { withCredentials: true })
+    .then(res => {
+      console.log('getUserInfoResponse', res);
+      return res;
+    })
+    .catch(err => {
+      console.log('getUserInfoError', err);
+      return err;
+    });
+  }
+  
+  function doesHttpOnlyCookieExist(cookiename) {
+    let date = new Date();
+    date.setTime(date.getTime() + (1000));
+    let expires = "expires=" + date.toUTCString();
+ 
+    document.cookie = cookiename + "=new_value;path=/;" + expires;
+    if (document.cookie.indexOf(cookiename + '=') == -1) {
+        return true;
+     } else {
+        return false;
+     }
+  }
+
+  let cookieExists = doesHttpOnlyCookieExist('connect.sid');
+  console.log(cookieExists);
+
+  // getUserInfo();
 
   return (
     <DropDownWrapper>
       {/* Conditional check to see if user is logged in */}
       {/* if not logged in, show the login/signup buttons */}
-      <LogInLink>Log In</LogInLink> <SignUpLink>Signup</SignUpLink>
       {/* if logged in, show component that says "Hello NAME then have a signout button" */}
+      { cookieExists ? ( 
+        <Fragment>
+        <WelcomeMessage>Welcome</WelcomeMessage> <SignOutLink>Signout</SignOutLink>
+        </Fragment>
+      )
+      : (
+        <Fragment>
+        <LogInLink>Log In</LogInLink> <SignUpLink>Signup</SignUpLink>
+        </Fragment>
+      )
+      }
     </DropDownWrapper>
   );
 }
