+++ server/routes/userRoutes.js
@@ -6,69 +6,10 @@ const ensureLoggedIn = require("connect-ensure-login").ensureLoggedIn(
 );
 const usersDB = require("../models/usersModel");
 
-const authenticate = require('../config/authMiddleware');
+// const authenticate = require('../config/authMiddleware');
 
-router.get(
-  "/signin",
-  passport.authenticate("auth0", {
-    scope: "openid email profile"
-  }),
-  function(req, res) {
-    res.redirect("/");
-  }
-);
-
-router.get("/callback", function(req, res, next) {
-  passport.authenticate("auth0", function(err, user, info) {
-    if (err) {
-      return next(err);
-    }
-    if (!user) {
-      return res.redirect("/signin");
-    }
-    req.logIn(user, function(err) {
-      if (err) {
-        return next(err);
-      }
-      const returnTo = req.session.returnTo;
-      delete req.session.returnTo;
-      // console.log('callback', req.user);
-      let role = req.user._json['https://ratemydiy.herokuapp.com/roles'];
-      console.log(role[0]);
-      let sub = req.user._json.sub.split("|");
-      let auth_id = sub[1];
-      let username = req.user._json.nickname;
-      let user = {
-        auth_id,
-        username
-      };
-      console.log(req.cookies);
-      if (role[0] === 'new') {
-        usersDB
-          .addUser(user)
-          .then(res => {
-              res.redirect(returnTo || "/");
-          })
-          .catch(err => {
-              res.status(500).json(err);
-          });
-      } else {
-        res.redirect(returnTo || "/");
-      }
-    });
-  })(req, res, next);
-});
-
-
-router.get('/signout', (req, res) => {
-	req.logout();
-	res.redirect('/');
-  });
-
-router.post("/test", ensureLoggedIn, authenticate, function(req, res, next) {
-  //console.log(req.user);
-  //console.log(req.user.app_metadata);
-  res.status(200).json({ message: 'it works' });
+router.get("/user", function(req, res, next) {
+  res.status(200).json(req.user);
 });
 
 module.exports = router;
