+++ server/server.js
@@ -25,13 +25,15 @@ server.get('/hello', (req, res) => {
 	res.send('hi');
 });
 
-const userRoutes = require('./routes/userRoutes');
+const authRoutes = require('./routes/authRoutes');
 const projectRoutes = require('./routes/projectRoutes');
 const postRoutes = require('./routes/postRoutes');
+const userRoutes = require('./routes/userRoutes');
 
-server.use('/api/users', userRoutes);
+server.use('/', authRoutes);
 server.use('/api/projects', projectRoutes);
 server.use('/api/posts', postRoutes);
+server.use('/api/users', userRoutes);
 
 // server.use(function(req, res, next) {
 //   res.header("Access-Control-Allow-Origin", "*");
